requirements.txt
tqdm
editdistance
pillow
opencv-python
torch
torchvision
tensorboardX
transformers
tensorflow-gpu==1.14.0
